Project restructure + homepage structuring
