+++ src/styles/page.module.css
@@ -8,7 +8,8 @@
 .header {
   display: flex;
   align-items: center;
-  position: relative;
+  position: sticky;
+  top: 0;
   padding: 10px 20px;
   background-color: #f5f5f5; /* Optional background color */
 }
